Move WiFi settings into rita_client
Also made common dashboard elements public, now we can have client
and server specific dashboard components easily.
