+++ rita/src/rita_client/mod.rs
@@ -1,3 +1,4 @@
 pub mod exit_manager;
 pub mod rita_loop;
 pub mod traffic_watcher;
+pub mod dashboard;
